[Support] Ensure operands are same type in configuration test file, NFC (#29590)
This silences a compiler warning because the operands to ASSERT_EQ were
not of the same type.

MODULAR_ORIG_COMMIT_REV_ID: 7de059ee519eadcfe18910df2d5ce91b1df5ca34

+++ Support/unittests/ConfigurationTest.cpp
@@ -459,16 +459,16 @@ key_3 = value3_override
   Config base;
   auto err = base.parseFrom(one);
   ASSERT_FALSE(err.isError()) << err.getError();
-  ASSERT_EQ(base.getAllValues().size(), 6);
+  ASSERT_EQ(base.getAllValues().size(), (size_t)6);
 
   Config over;
   err = over.parseFrom(two);
   ASSERT_FALSE(err.isError()) << err.getError();
-  ASSERT_EQ(over.getAllValues().size(), 5);
+  ASSERT_EQ(over.getAllValues().size(), (size_t)5);
 
   err = base.overrideFrom(over);
   ASSERT_FALSE(err.isError()) << err.getError();
-  ASSERT_EQ(base.getAllValues().size(), 8);
+  ASSERT_EQ(base.getAllValues().size(), (size_t)8);
 
   EXPECT_EQ(base.getValue("global_key_1"), "value1");
   EXPECT_EQ(base.getValue("global_key_2"), "value2_override");
